requirements.txt
streamlit
langgraph
langchain-google-genai
pydantic
python-dotenv
PyPDF2